cleanup google auth
