+++ packages/medusa-plugin-auth/src/api/index.ts
@@ -3,13 +3,13 @@ import { ConfigModule } from '@medusajs/medusa/dist/types/global';
 import loadConfig from '@medusajs/medusa/dist/loaders/config';
 
 import { AuthOptions } from '../types';
+import { loadJwtOverrideStrategy } from '../auth-strategies/jwt-override';
 import { getGoogleAdminAuthRouter, getGoogleStoreAuthRouter } from '../auth-strategies/google';
-import { loadJwtOverrideStrategy } from "../auth-strategies/jwt-override";
 
 export default function (rootDirectory, pluginOptions: AuthOptions): Router[] {
 	const configModule = loadConfig(rootDirectory) as ConfigModule;
 
-	loadJwtOverrideStrategy(configModule)
+	loadJwtOverrideStrategy(configModule);
 
 	return loadRouters(configModule, pluginOptions);
 }
