+++ packages/medusa-plugin-auth/src/auth-strategies/google/index.ts
@@ -1,0 +1,4 @@
+export const ENTITY_METADATA_KEY = 'useGoogleStrategy';
+
+export * from './admin';
+export * from './store';
