+++ packages/medusa-plugin-auth/src/loaders/index.ts
@@ -1,7 +1,7 @@
 import { ConfigModule, MedusaContainer } from '@medusajs/medusa/dist/types/global';
-import { loadGoogleAdminStrategy, loadGoogleStoreStrategy } from '../auth-strategies/google';
 
 import { AuthOptions } from '../types';
+import { loadGoogleAdminStrategy, loadGoogleStoreStrategy } from '../auth-strategies/google';
 
 export default async function authStrategiesLoader(container: MedusaContainer, authOptions: AuthOptions) {
 	const configModule = container.resolve('configModule') as ConfigModule;
