+++ packages/medusa-plugin-auth/src/types/index.ts
@@ -1,4 +1,4 @@
-export const AUTH_TOKEN_COOKIE_NAME = "auth_token";
+export const AUTH_TOKEN_COOKIE_NAME = 'auth_token';
 
 export type AuthOptions = {
 	google?: {
