+++ packages/medusa-plugin-auth/src/utils/get-cookie-options.ts
@@ -1,0 +1,17 @@
+import { CookieOptions } from 'express-serve-static-core';
+
+export function getCookieOptions(): CookieOptions {
+	let secure = false;
+	let sameSite: CookieOptions['sameSite'] = false;
+
+	if (process.env.NODE_ENV === 'production' || process.env.NODE_ENV === 'staging') {
+		secure = true;
+		sameSite = 'none';
+	}
+
+	return {
+		sameSite,
+		secure,
+		maxAge: 10 * 60 * 60 * 1000,
+	};
+}
